requirements.txt
asgiref==3.7.2
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.3.0
Django==5.0
django-channels-jwt==0.0.3
django-cors-headers==4.4.0
django-csp==3.7
django-jet-reboot==1.3.7
django-js-asset==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
redis==5.1.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
uvicorn[standard]
django-redis==5.4.0
whitenoise